Refactored controllers, added helpers module
