+++ app/controllers/sessions.rb
@@ -5,25 +5,23 @@ class Beast < Sinatra::Base
 	end
 
 	get '/sessions/new' do
-	 	erb :'sessions/new', :layout => false
+		erb :'sessions/new', :layout => false
 	end
 
 	post '/sessions' do
-	  user = User.authenticate(params[:email], params[:password])
-	  if user
-			@spaces = Space.all
-	  	session[:user_id] = user.id
-	    redirect to('/spaces/all')
-	  else
-	    flash.now[:errors] = ['The email or password is incorrect']
-	    erb :'sessions/new'
-	  end
+		user = User.authenticate(params[:email], params[:password])
+		if user
+			session[:user_id] = user.id
+			redirect to('/spaces/all')
+		else
+			flash.now[:errors] = ['The email or password is incorrect']
+			erb :'sessions/new'
+		end
 	end
 
 	delete '/sessions' do
-		@spaces = Space.all
-	  session[:user_id] = nil
-	  flash.keep[:notice] = 'goodbye!'
-	  redirect to '/spaces/all'
+		session[:user_id] = nil
+		flash.keep[:notice] = 'goodbye!'
+		redirect to '/spaces/all'
 	end
 end
