+++ app/server.rb
@@ -1,16 +1,13 @@
 class Beast < Sinatra::Base
-
   register Sinatra::Flash
   register Sinatra::Partial
   use Rack::MethodOverride
+
+  helpers Helpers
+
   set :partial_template_engine, :erb
   enable :partial_underscores
   enable :sessions
   set :session_secret, 'super secret'
 
-  helpers do
-    def current_user
-      @current_user ||= User.get(session[:user_id])
-    end
-  end
 end
