+++ app/views/spaces/confirmations/booked.erb
@@ -1,4 +1,3 @@
 Request approved
 
-<%= p @space.available_dates %>
-<%= p @space.available_dates.first %>
+<%= @space.available_dates %>
