+++ app/views/users/new.erb
@@ -25,17 +25,17 @@
               <label class="sr-only" for="name">Name</label>
               <input name="name" type="text" placeholder="Name..." value="<%=@user.name%>" class="form-control" required >
             </div>
-            
+
             <div class="form-group">
               <label class="sr-only" for="email">Email</label>
               <input name="email" type="email" placeholder="Email..." value="<%=@user.email%>" class="form-control" required>
             </div>
-            
+
             <div class="form-group">
               <label class="sr-only" for="password">Password</label>
               <input name="password" type="password" placeholder="Password..." class="form-control" required>
             </div>
-            
+
             <div class="form-group">
               <label class="sr-only" for="password_confirmation">Confirm Password</label>
               <input name="password_confirmation" type="password" placeholder="Confirm password..." class="form-control" required>
